Added HttpResponse to the RestControllers

+++ src/main/java/com/urjc/daw/practica/api/QuoteRestController.java
@@ -4,7 +4,11 @@ package com.urjc.daw.practica.api;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.support.PagedListHolder;
+import org.springframework.data.domain.Page;
+import org.springframework.expression.spel.ast.OpInc;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,6 +17,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
@@ -20,7 +25,7 @@ import com.urjc.daw.practica.model.Quote;
 import com.urjc.daw.practica.service.impl.QuoteManagementServiceImpl;
 
 @RestController
-@RequestMapping("/api/quote")
+@RequestMapping("/api/quotes")
 public class QuoteRestController {
 
 	private static final int QUOTES_PER_PAGE = 10;
@@ -29,18 +34,23 @@ public class QuoteRestController {
 	QuoteManagementServiceImpl quoteService;
 
 	@GetMapping("/{id}")
-	public Optional<Quote> findOne(@PathVariable long id) {
-		return quoteService.findOne(id);
+	public ResponseEntity<Optional<Quote>> findOne(@PathVariable long id) {
+		Optional<Quote> found = quoteService.findOne(id);
+		if(found.isPresent()) {
+			return new ResponseEntity<>(found,HttpStatus.OK);
+		}else {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
 	}
 
-	@GetMapping("/all/{nPag}")
-	public List<Quote> findAll(@PathVariable int nPag) {
-		return quoteService.findAll(nPag, QUOTES_PER_PAGE);
-	}
-	
-	@GetMapping("/all")
-	public List<Quote> findAll(){
-		return quoteService.findAll(1, QUOTES_PER_PAGE);
+	@GetMapping(params = {"page"}) 
+	public ResponseEntity<Page<Quote>> findAll(@RequestParam("page") int nPag) {
+		Page<Quote> pageable = quoteService.findAll(nPag, QUOTES_PER_PAGE);
+		if(pageable.hasContent()) {
+			return new ResponseEntity<>(pageable,HttpStatus.OK);
+		}else {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
 	}
 
 	@PostMapping("/")
@@ -52,18 +62,26 @@ public class QuoteRestController {
 	}
 
 	@PutMapping("/{id}")
-	public Quote editQuote(@PathVariable long id,@RequestBody Quote quote) {
-		quoteService.findOne(id).get();
-		quote.setId(id);
-		quoteService.save(quote);
-		return quote;
+	public ResponseEntity<Quote> editQuote(@PathVariable long id,@RequestBody Quote quote) {
+		Quote found = quoteService.findOne(id).get();
+		if(found != null) {
+			quote.setId(id);
+			quoteService.save(quote);
+			return new ResponseEntity<>(found,HttpStatus.OK);
+		}else {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
 	}
 
 	@DeleteMapping("/{id}")
-	public Optional<Quote> deleteQuote(@PathVariable long id) {
+	public ResponseEntity<Optional<Quote>> deleteQuote(@PathVariable long id) {
 		Optional<Quote> deleted = quoteService.findOne(id);
-		quoteService.deleteQuote(id);
-		return deleted;
+		if(deleted.isPresent()) {
+			quoteService.deleteQuote(id);
+			return new ResponseEntity<>(deleted,HttpStatus.OK);
+		}else {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
 	}
 
 	@GetMapping("/")
